add: replace "mobileBreakpoint()" to "defaultBreakpoint"
add: replace "mobileBreakpoint()" to "defaultBreakpoint"

sometimes it is hard to find the css error when there are no caller syntax "()" when using "mobileBreakpoint".

it is designed to provide customizable breakpoint syntax but only used a single default breakpoint.

so then created not a function but a string value for default breakpoint syntax.